Use modules to refactor away from a single large Javascript file.
