+++ index.html
@@ -38,7 +38,8 @@ Date: Spring 2023
 }
     </style>
 
-    <script src="videoManagement.js"></script>
+    <script src="scripts/newVideo.js" type="module"></script>
+    <script src="scripts/videoList.js" type="module"></script>
 
 </head>
 <body onload="fillVideoList()">
